Install steam from neopkgs
I had to modify neopkgs to pass arguments to the version of nixpkgs it
imports so I can configure it to allow proprietary software with:

    config.allowUnfree = true;

This seems to fix some things with running my games with steam.  I can
actually run Borderlands 2 now (before it would die before anything
visible happened).  Now, Borderlands 2 will play the intro, but it dies
after it checks for DLC and updates.