Refactor out common.nix
With the release of NixOS 16.09 I've realized that I want to be able to
upgrade each machine individually.  So I define `system.stateVersion` in
each machine's configuration.

I also moved `time.timeZone` to each machine's configuration.  They're
all the same for the moment, but it makes sense to keep it in the
machine definition (in the event that I take a machine with me when I
travel).

The definition to install git in `common.nix` is redundant with
`workstation.nix` so I removed `common.nix` entirely.

Now, there's no one configuration that ties all machines together, each
machine can import a specific role, but there's no longer a
configuration that all machines "inherit" from.